tools/requirements.txt
bleak==0.22.1
fire==0.6.0
Pillow==10.3.0
Requests==2.31.0
rich==13.7.1
